server/datacenter/bin/controller/code/requirements.txt
configparser
psycopg2-binary